+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Names of the foreign objects in the Records table are displayed next to their IDs.
+  ([#570](https://github.com/TheRacetrack/racetrack/issues/570))
 
 ## [2.35.0] - 2025-02-05
 ### Added